Add LocalFunction::bound() to internat type erasure class and test
Surprisingly the test worked before, although it does a concept check
for the interface. The reason is, that the interface of the type erasure
class contains the `bound()` method, so the substitution in the concept
check works. But this just checks if the expression is valid when
instantiating the template. It does not check if the instantiated
template compiles. Hence we really need to call the method in the check.