Test power of a ReducedHermite basis
This new test currently leads to a segfault, because somewhere memory
gets corrupted when binding the leaf nodes to particular elements.
I have yet to fix this.

To publish my current knowledge, this patch contains a simple call
sequence that I extracted from the general test that makes the code
crash.  Binding a localView and then asking the local basis for the
element type is enough.