Add support for Hermite elements to the global interpolate() method
Currently this contains two cases (activated via a macro):

* (A) Local interpolation of `fLocal` assumes that `derivative(fLocal)`
  is really the derivative of `fLocal`.
* (B) Local interpolation of `fLocal` assumes that `derivative(fLocal)`
  is the localized version of the global derivative.

See https://gitlab.dune-project.org/staging/dune-functions/-/issues/76
for a discussion of the two variants.

This also included support for local functions `lf` where `derivative(lf)`
hands out a reference to a cached derivative. This potentially allows
to avoid the creation and copy of those local derivatives on each element
which can be relatively costly as e.g. in case of `DiscreteGlobalBasisFunction`.
However, the latter does not implement such caching so far but could easily
be patched to provide it.