[bug][draft] Switch poisson-pq2.py to triangle grid
This implements the following changes:

* Add a gmsh coarse grid file consisting of two triangles
  on the unit square.
* Load and refine the grid in `poisson-pq2.py`.
* Add some debug output to `poisson-pq2.py`.

When using the grid, the example fails with a segfault.
This does not come from the assembler, but from the
boundary treatment code that is generated from C++.
Adding debug output reveals that here the basis
provides buggy global indices. In fact even the
number of elements is not correct here.