[nodeTree] tag dispatch based implementation of bindTree
in order to implement functionspacebases which are stricted to a
subdomain, we need to support additional nodes with very specific
behaviour. This can not achieved with the TypeTree visitor
pattern. This is a reimplementation of the bind without using a
visitor, but a tag-dispatch. That allows to add the special node
handling strategies in the future, e.g. skip the bind for nodes which
do not have support in a given subdomain.