[draft] Implement higher order Lagrange basis in 3d
This implements DOF reordering for faces required for higher order
Lagrange bases in 3d. To guarantee continuity across faces of all
codimensions for bases in 2d and 3d, we need to reorder DOFs on
face subentities of type line/triangle/quadrilateral. For edges
this was already implemented, This patch implements the same for
triangles and quadrilateral and unifies the implementation:

* We now use an `MCMGMapper` instead of computing all offsets by hand.
  This simplifies the code significantly.
* For all three types of faces we introduce an orientation represented
  by a number. The bits of the number indicate the required subentity
  reflections.
* The reordering of triangle/quadrilateral face DOFs is precomputed
  in a table for each orientation.
* When computing indices we first compute the orientation for the
  face subentity and derive the DOF reordering form the table.

In principle this should also work for higher order pyramid and prism
elements as long as face DOFs are ordered canonically on the reference
element. However, the compile-time order implementation is limited to
order=1,2 and the run-time order implementation is broken even for
order=1 (independently of this patch).

While the implementation is fully functional, this is currently
marked as a draft, because several things are missing so far:

* Adjust the basis test to carefully test all combinations with
  different dynamic/static order. For higher _dynamic_ order this
  requires to relax the test tolerances.
* Update documentation on implemented cases.
* Add exceptions for non-implemenented cases (dim=3, higher order pyramid
  elements for dim=3)
* Do some benchmarking to ensure that the code is not slower for the
  simple cases.
* Maybe precompute the subentity orientations instead of recomputing
  them for each DOF.