Allow Lagrange basis to be bound to entities outside of its domain
The grid view transformation of basis and pre-basis nodes may be
bound to regions where the underlying functions are not defined, so its size is set zero. If computations on nodes with zero size are skipped, undefined finite elements are also skipped. This allows grid-view transformations to oversample the domain without  triggering undefined behavior.