Introduce the SubDomain class representing a subset of a grid view
The class allows to create a grid view which implements an index set,
a `contains()` method, element- and intersection-iterators.

If we have some experrience with this, it miht be a candidate
for beeing moved to dune-grid.

This is inspired by the `MultiDomainGridView` proposed by Christian Engwer
in https://gitlab.dune-project.org/staging/dune-functions/-/merge_requests/494
but strips done the implementation to a more light weight approach.