[subdomain] Add SubDomainInterface class
This represents the interface between two subdomains
and conceptually represents a range of intersections.
It allows to iterate over the intersections and provides
a contains method to check if an intersection is between
the subdomains.

Notice that the iterator will visit intersections only
once with a well defined orientation while the `contains()`
method does not care for the orientation.