Throw exception if multi-index does not match nested container
If the multi-index is exhausted but we did not terminate the
container access recursion we now throw and exception instead
of implicitly pretending that there are more zeros in the
multi-index. As a conseqeunce the vector backend no longer treats
`Dune::FieldVector<K,1>` like a scalar `K`.

Notice that until recently even own examples made use of the
implicit padding by zeros.